Time: 0 ms (100%), Space: 10.7 MB (10.38%) - LeetHub

+++ 0410-split-array-largest-sum/0410-split-array-largest-sum.cpp
@@ -1,17 +1,17 @@
 class Solution {
 public:
     int splitArray(vector<int>& nums, int k) {
-        long long l = *max_element(nums.begin() , nums.end());
-        long long h = accumulate(nums.begin() , nums.end() , 0LL);
+        int l = *max_element(nums.begin() , nums.end());
+        int h = accumulate(nums.begin() , nums.end() , 0LL);
 
-        long long ans = -1;
+        int ans = -1;
 
         while(l <= h){
 
-            long long m = l + (h - l)/2;
+            int m = l + (h - l)/2;
 
-            long long sum = 0;
-            long long subarray = 1;
+            int sum = 0;
+            int subarray = 1;
 
             for(auto i : nums){
                 if(sum + i > m){
